#416 make the switch in a single configuration
